Only drop bucket if it exists

+++ api/couchbase-scripts/drop-db.sh
@@ -7,5 +7,8 @@ if [ $# -eq 1 ]; then
     bucket="$1"
 fi
 
-curl -X DELETE -u poleland:poleland \
-    http://10.0.0.2:8091/pools/default/buckets/$bucket
+existsStatusCode=$(curl -s -o /dev/null -w "%{http_code}" http://10.0.0.2:8091/pools/default/buckets/$bucket)
+if [ $existsStatusCode -eq 200 ]; then
+    curl -X DELETE -u poleland:poleland \
+        http://10.0.0.2:8091/pools/default/buckets/$bucket
+fi
